[tests/tlgamma.c] Improved a test in mpfr_lgamma_nosign.
Since we test sign != 1 && sign != -1 after a call with &sign, we first
initialize sign to 0 (a value different from 1 and -1) so that we can
detect when it is not explicitly set to 1 or -1 in mpfr_lgamma.

A failure was appearing with -O2, but not with -O3 (because the
uninitialized sign happened to be 1 or -1 during the execution).
With this change, the failure appears at all optimization levels.