[mpfr.texi] Moved the note about underflow and IEEE 754, and update.
Moved the note about underflow and IEEE 754 to the description of
mpfr_subnormalize, because only this function is concerned. Indeed,
when subnormals are not supported (which is the case of MPFR), an
underflow case is necessarily inexact.

In case of underflow, the mpfr_subnormalize function also sets the
underflow flag when the result is exact:

  https://sympa.inria.fr/sympa/arc/mpfr/2009-06/msg00010.html

Hence the note.