Portability: clean up support for binary128 (a.k.a. float128).
This fixes a configure failure with Clang and glibc 2.41:
  https://bugs.gentoo.org/949962
  https://sourceware.org/bugzilla/show_bug.cgi?id=32750
  https://sympa.inria.fr/sympa/arc/mpfr/2025-02/msg00025.html
  https://sympa.inria.fr/sympa/arc/mpfr/2025-03/msg00001.html

The issue is the following: in its configure script, GNU MPFR first
tests whether _Float128 is supported (without header inclusions);
if it isn't, MPFR then tests whether __float128 is supported, and
in such a case, it defines _Float128 as __float128 with

  AC_DEFINE([_Float128],[__float128],[__float128 fallback])

As Clang supports __float128 but not _Float128 (at least on x86_64,
up to version 19), this fallback is selected. But glibc 2.41 has a
similar fallback, with a typedef:

  typedef __float128 _Float128;

With the MPFR _Float128 macro defined as __float128, as soon as some
header such as <stdio.h> is included, this yields

  typedef __float128 __float128;

which is incorrect. This first occurs in the

  AC_MSG_CHECKING(for GMP_NUMB_BITS and sizeof(mp_limb_t) consistency)
  [...]

configure test (in configure.ac), hence the failure of this test and
the configure script.

As an attempt to avoid such an issue with header inclusion, a solution
should be to include the common headers in every configure test.

But since _Float128 is a reserved identifier, we must avoid defining
it as a macro (a typedef would also be incorrect). This is done here,
by defining a mpfr_float128 macro instead. In the public mpfr.h header
file, if the user has defined MPFR_WANT_FLOAT128: mpfr_float128 will
be defined as _Float128 by default (thus this will not change the API),
but if _Float128 is not supported, the user should define mpfr_float128
as __float128 (or an equivalent type that is supported).

* acinclude.m4: changed _Float128 to mpfr_float128 in AC_DEFINE.
* configure.ac: updated a comment about _Float128.
* doc/README.dev: update about MPFR_WANT_FLOAT128 and _Float128.
* doc/mpfr.texi: in "Nomenclature and Types", added a paragraph about
  binary128 support and the mpfr_float128 macro; update concerning the
  functions related to binary128 (in particular, in the prototypes,
  changed _Float128 to mpfr_float128).
* src/mpfr.h: under "#ifdef MPFR_WANT_FLOAT128", define mpfr_float128
  as _Float128 if not defined yet; in the prototypes, changed
  _Float128 to mpfr_float128.
* src/get_float128.c, src/set_float128.c, tests/tset_float128.c:
  changed _Float128 to mpfr_float128; updated comments.
* tests/tversion.c: for "float128 = yes", also output the actual type.