[tests/tversion.c] Do not output gmp_*printf info when meaningless.
In particular, mini-gmp does not provide gmp_*printf functions, thus
it does not make any sense to output the corresponding lines.

Until now, one was getting

[tversion] gmp_printf: hhd = yes, lld = yes, jd = yes, td = ?, Lf = ?
[tversion] gmp_snprintf/vsnprintf on "%a" = ?

The "?" corresponded to the absence of the tests, thus could be
regarded as correct. But the "yes" was just the default, thus was
actually incorrect and misleading. Note that this default does not
matter for mini-gmp since the formatted output functions are not
tested: the corresponding test files have

  #if defined(HAVE_STDARG) && !defined(MPFR_USE_MINI_GMP)

and we are now using the same test.